+++ src/interfaces/news.ts
